requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cssselect==1.3.0
elevenlabs==0.3.0
fastapi==0.109.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
google-generativeai==0.3.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.10
importlib_metadata==7.2.1
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml[html_clean]==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.38.2
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
packaging==23.2
pandas==2.2.3
pillow==10.4.0
proto-plus==1.26.1
protobuf==4.25.7
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.4.0
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==80.4.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sseclient-py==1.8.0
starlette==0.36.3
streamlit==1.32.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.0
torchvision==0.17.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.40.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.27.1
validators==0.35.0
zipp==3.21.0
